+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.20] - 2026-06-04
+
 ### Fixed
 
 - `sla_ancestors_100k` macOS CI flake: bumped macOS-specific SLA limit from